refactor: move `construct_search_params` to shared `search` utility module
